pyproject.toml
[project]
name = "example-worker-load-balancing-py"
version = "0.1.0"
description = "Service discovery, load balancing, and recovery example application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "resonate-sdk>=0.6.0",
]
